Secure views-Product admin

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
 
 from .models import Product, Category
@@ -69,12 +70,15 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     """
     A view for add product to the store
     """
-   
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not authorized to add products!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -87,7 +91,6 @@ def add_product(request):
             messages.error(
                 request, 'Failed to add the new product. \
                     Please ensure the form is valid.')
-           
     else:
         form = ProductForm()
 
@@ -98,11 +101,15 @@ def add_product(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """
     A view for store manager to edit selected products
     """
-    
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not authorized to add products!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -113,9 +120,7 @@ def edit_product(request, product_id):
         else:
             messages.error(request, 'Failed to update the product. \
                 Please ensure the form is valid.')
-          
-           
-           
+
     else:
         form = ProductForm(instance=product)
         messages.info(
@@ -129,12 +134,14 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
     """
     A view for store managers to delete products
     """
-   
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not authorized to add products!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
